Add gperf to the analysis
While I already have a way to make a perfect hash from the contents of
the weather stations, the hash is a very large value. The one from gperf
is smaller and it could be interesting to see if it could be used for a
sparse map instead of needing to modulo and check for collisions on the
modulo.